created new folders & moved page test to correct folder
